requirements.txt
python-dotenv
psycopg2-binary
numpy
